Move fields to their form components
